backend/pyproject.toml
[project]
name = "fit-tracker-backend"
version = "0.1.0"
description = "FastAPI backend for integrations, metrics, syncs, and insights."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "fastapi>=0.116.0,<1.0.0",
  "greenlet>=3.2.0,<4.0.0",
  "httpx>=0.28.0,<1.0.0",
  "pandas>=2.3.0,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "python-dotenv>=1.1.0,<2.0.0",
  "sqlalchemy>=2.0.43,<3.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "apscheduler>=3.11.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.17.0,<2.0.0",
  "pytest>=8.4.0,<9.0.0",
  "pytest-asyncio>=1.1.0,<2.0.0",
  "ruff>=0.12.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*", "tests*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "ASYNC"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
